+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
   - Update braintree-web to v3.85.2
+  - Add support for Elo, Hiper, and Hipercard
   - Add support for UnionPay
 
 ## 1.32.1